requirements.txt
openai
chromadb
pyyaml
langchain
langchain-openai
tiktoken
matplotlib
scikit-learn
# Version: 1.1
# Updated: 2025-05-12 16:50:46
# Version: 1.2
# Updated: 2025-05-12 17:10:00 